[#17] BookMark,Follow 구현완료 최종

+++ src/main/java/com/example/capstone1/api/bookMark/repository/BookMarkRepository.java
@@ -9,16 +9,18 @@ import org.springframework.data.jpa.repository.Query;
 import javax.transaction.Transactional;
 import java.util.List;
 
-public interface BookMarkRepository extends JpaRepository<BookMark, Integer> {
+public interface BookMarkRepository extends JpaRepository<BookMark, Long> {
 
     @Query("select b from BookMark b where b.users.id=:UsersId")
     List<BookMark> findBookMarkByUsersId(Long UsersId);
 
-
+/*
     @Modifying
     @Transactional
     @Query("delete from BookMark bookmark where bookmark.id=:bookMarkId")
     void deleteBookMarkById(long bookMarkId);
+
+ */
 
 
 }
