+++ src/main/java/com/example/capstone1/api/follow/entity/Follow.java
@@ -16,15 +16,15 @@ import javax.persistence.*;
 public class Follow extends BaseTime { //좋아요
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id; //팔로우 식별 아이디
+    private long id; //팔로우 식별 아이디
 
     @Column
     private String Userid; //팔로우 하는 사람
-/*
+
     @ManyToOne
     @JoinColumn(name = "USERS_ID")
     private Users users; //팔로우 된 사람
 
- */
+
 
 }
